Add files via upload
As a data analyst you should be able to work as a team and work effectively with a team to achieve a positive goal.
Here is practical example of a team work that was presented and we are rated amidst the top three out of 16 group.  We could visualize your data and analyze them for you to know were more efforts are needed, to attain your project aim.